+++ src/datasources/results/query-handlers/query-steps/QueryStepsDataSource.ts
@@ -1,7 +1,6 @@
 import { DataQueryRequest, DataFrameDTO, FieldType, LegacyMetricFindQueryOptions, MetricFindValue } from '@grafana/data';
 import { OutputType } from 'datasources/results/types/types';
 import {
-  QueryResponse,
   QuerySteps,
   QueryStepsResponse,
   StepsProperties,
@@ -15,6 +14,8 @@ import { StepsQueryBuilderFieldNames } from 'datasources/results/constants/Steps
 import { ExpressionTransformFunction, transformComputedFieldsQuery } from 'core/query-builder.utils';
 import { ResultsQueryBuilderFieldNames } from 'datasources/results/constants/ResultsQueryBuilder.constants';
 import { StepsVariableQuery } from 'datasources/results/types/QueryResults.types';
+import { QueryResponse } from 'core/types';
+import { queryInBatches } from 'core/utils';
 
 export class QueryStepsDataSource extends ResultsDataSourceBase {
   queryStepsUrl = this.baseUrl + '/v2/query-steps';
@@ -83,7 +84,7 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
       requestsPerSecond: QUERY_STEPS_REQUEST_PER_SECOND
     };
 
-    const response = await this.queryInBatches(queryRecord, batchQueryConfig, take);
+    const response = await queryInBatches(queryRecord, batchQueryConfig, take);
 
     return {
       steps: response.data,
@@ -91,7 +92,7 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
       totalCount: response.totalCount
     };
   }
-  
+
   async runQuery(query: QuerySteps, options: DataQueryRequest): Promise<DataFrameDTO> {
     if (!query.resultsQuery) {
       return {
@@ -99,7 +100,7 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
         fields: [],
       };
     }
-    
+
     query.stepsQuery = this.transformQuery(query.stepsQuery, this.stepsComputedDataFields, options);
     query.resultsQuery = this.transformQuery(query.resultsQuery, this.resultsComputedDataFields, options) || '';
 
@@ -108,7 +109,7 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
     const projection = query.showMeasurements
       ? [...new Set([...(query.properties || []), StepsPropertiesOptions.DATA])]
       : query.properties;
-    
+
     if (query.outputType === OutputType.Data) {
       const responseData = await this.queryStepsInBatches(
         stepsQuery,
@@ -119,7 +120,7 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
         query.resultsQuery,
         true
       );
-  
+
       if (responseData.steps.length === 0) {
         return {
           refId: query.refId,
@@ -247,16 +248,16 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
    * @returns - The transformed query string, or undefined if the input queryField is undefined.
    */
   private transformQuery(queryField: string | undefined, computedDataFields: Map<string, ExpressionTransformFunction>, options: DataQueryRequest): string | undefined {
-  return queryField
-    ? transformComputedFieldsQuery(
+    return queryField
+      ? transformComputedFieldsQuery(
         this.templateSrv.replace(queryField, options.scopedVars),
         computedDataFields
       )
-    : undefined;
+      : undefined;
   }
 
   async metricFindQuery(query: StepsVariableQuery, options?: LegacyMetricFindQueryOptions): Promise<MetricFindValue[]> {
-    if (query.queryByResults !== undefined && this.isTakeValid(query.take!)) {
+    if (query.queryByResults !== undefined && this.isTakeValid(query.stepsTake!)) {
       const resultsQuery = query.queryByResults ? transformComputedFieldsQuery(
         this.templateSrv.replace(query.queryByResults, options?.scopedVars),
         this.resultsComputedDataFields
@@ -273,7 +274,7 @@ export class QueryStepsDataSource extends ResultsDataSourceBase {
           stepsQuery,
           'UPDATED_AT',
           [StepsPropertiesOptions.NAME as StepsProperties],
-          query.take,
+          query.stepsTake,
           true,
           resultsQuery,
         );