+++ src/datasources/work-orders/components/WorkOrdersQueryEditor.tsx
@@ -1,7 +1,7 @@
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { QueryEditorProps, SelectableValue } from '@grafana/data';
 import { WorkOrdersDataSource } from '../WorkOrdersDataSource';
-import { OutputType, WorkOrderProperties, WorkOrderPropertiesOptions, WorkOrdersQuery } from '../types';
+import { OrderBy, OutputType, WorkOrderProperties, WorkOrderPropertiesOptions, WorkOrdersQuery } from '../types';
 import { WorkOrdersQueryBuilder } from './query-builder/WorkOrdersQueryBuilder';
 import {
   AutoSizeInput,
@@ -11,7 +11,6 @@ import {
   Select,
   VerticalGroup
 } from '@grafana/ui';
-import { OrderBy } from 'datasources/products/types';
 import './WorkOrdersQueryEditor.scss';
 import { validateNumericInput } from 'core/utils';
 import { TAKE_LIMIT } from 'datasources/results/constants/QuerySteps.constants';
@@ -21,6 +20,11 @@ type Props = QueryEditorProps<WorkOrdersDataSource, WorkOrdersQuery>;
 export function WorkOrdersQueryEditor({ query, onChange, onRunQuery, datasource }: Props) {
   query = datasource.prepareQuery(query);
   const [recordCountInvalidMessage, setRecordCountInvalidMessage] = useState<string>('');
+
+  useEffect(() => {
+    handleQueryChange(query, true);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []); // Only run on mount
 
   const handleQueryChange = useCallback(
     (query: WorkOrdersQuery, runQuery = true): void => {
@@ -47,6 +51,13 @@ export function WorkOrdersQueryEditor({ query, onChange, onRunQuery, datasource 
 
   const onDescendingChange = (isDescendingChecked: boolean) => {
     handleQueryChange({ ...query, descending: isDescendingChecked });
+  };
+
+  const onQueryByChange = (queryBy: string) => {
+    if(query.queryBy !== queryBy) {
+      query.queryBy = queryBy;
+      handleQueryChange({ ...query, queryBy });
+    }
   };
 
   const onTakeChange = (event: React.FormEvent<HTMLInputElement>) => {
@@ -98,7 +109,11 @@ export function WorkOrdersQueryEditor({ query, onChange, onRunQuery, datasource 
           </InlineField>
         )}
         <InlineField label="Query By" labelWidth={25} tooltip={tooltips.queryBy}>
-            <WorkOrdersQueryBuilder globalVariableOptions={[]}></WorkOrdersQueryBuilder>
+            <WorkOrdersQueryBuilder
+              filter={query.queryBy} 
+              globalVariableOptions={[]}
+              onChange={(event: any) => onQueryByChange(event.detail.linq)}
+            ></WorkOrdersQueryBuilder>
           </InlineField>
         </VerticalGroup>
         <VerticalGroup>